(feat): Added the whole frontend.

+++ src/Events.IO.Web/Controllers/EventsController.cs
@@ -22,6 +22,11 @@ namespace Events.IO.Site.Controllers
         {
             return View(_eventAppService.GetAll());
         }
+        [Authorize] 
+        public IActionResult MyEvents()
+        {
+            return View(_eventAppService.GetEventByHost(HostId));
+        }
 
         public IActionResult Details(Guid? id)
         {
@@ -48,7 +53,7 @@ namespace Events.IO.Site.Controllers
         [ValidateAntiForgeryToken]
         public IActionResult Create(EventViewModel eventViewModel)
         {
-            //if (!ModelState.IsValid) return View(eventViewModel);
+            if (!ModelState.IsValid) return View(eventViewModel);
 
             eventViewModel.HostId = HostId;
             _eventAppService.Registry(eventViewModel);
@@ -66,6 +71,15 @@ namespace Events.IO.Site.Controllers
             }
 
             var eventViewModel = _eventAppService.GetById(id.Value);
+
+            if (eventViewModel == null)
+            {
+                return NotFound();
+            }
+            if (ValidateAuthorEvent(eventViewModel))
+            {
+                return RedirectToAction("MyEvents", _eventAppService.GetEventByHost(HostId));
+            }
             if (eventViewModel == null)
             {
                 return NotFound();
@@ -77,13 +91,25 @@ namespace Events.IO.Site.Controllers
         [ValidateAntiForgeryToken]
         public IActionResult Edit(EventViewModel eventViewModel)
         {
-
+            if (ValidateAuthorEvent(eventViewModel))
+            {
+                return RedirectToAction("MyEvents", _eventAppService.GetEventByHost(HostId));
+            }
             if (!ModelState.IsValid) return View(eventViewModel);
 
+            eventViewModel.HostId = HostId;
             _eventAppService.Update(eventViewModel);
 
             ViewBag.PostReturn = ValidOperation() ? "success,Event updated!" : "error,Event was not updated verify the messages!";
 
+            if (_eventAppService.GetById(eventViewModel.Id).Online) {
+                eventViewModel.Address = null;
+            }
+            else
+            {
+                eventViewModel = _eventAppService.GetById(eventViewModel.Id);
+            }
+                
 
             return View(eventViewModel);
         }
@@ -96,6 +122,11 @@ namespace Events.IO.Site.Controllers
             }
 
             var eventViewModel = _eventAppService.GetById(id.Value);
+            
+            if (ValidateAuthorEvent(eventViewModel))
+            {
+                return RedirectToAction("MyEvents", _eventAppService.GetEventByHost(HostId));
+            }
 
             if (eventViewModel == null)
             {
@@ -109,6 +140,10 @@ namespace Events.IO.Site.Controllers
         [ValidateAntiForgeryToken]
         public IActionResult DeleteConfirmed(Guid id)
         {
+            if (ValidateAuthorEvent(_eventAppService.GetById(id)))
+            {
+                return RedirectToAction("MyEvents", _eventAppService.GetEventByHost(HostId));
+            }
             _eventAppService.Delete(id);
             return RedirectToAction("Index");
         }
@@ -120,6 +155,52 @@ namespace Events.IO.Site.Controllers
             }
             var eventViewModel = _eventAppService.GetById(id.Value);
             return PartialView("_IncludeAddress", eventViewModel);
+        }
+        public IActionResult UpdateAddress(Guid? id)
+        {
+            if (id == null)
+            {
+                return NotFound();
+            }
+            var eventViewModel = _eventAppService.GetById(id.Value);
+            return PartialView("_UpdateAddress", eventViewModel);
+        }
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public IActionResult IncludeAddress(EventViewModel eventViewModel)
+        {
+            ModelState.Clear();
+            eventViewModel.Address.EventId = eventViewModel.Id;
+            _eventAppService.AddAddress(eventViewModel.Address);
+
+            if (ValidOperation())
+            {
+                string url = Url.Action("GetAddress", "Events", new { id = eventViewModel.Id });
+                return Json(new { success = true, url = url });
+            }
+            return PartialView("_IncludeAddress", eventViewModel);
+        }
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public IActionResult UpdateAddress(EventViewModel eventViewModel)
+        {
+            ModelState.Clear();
+            _eventAppService.UpdateAddress(eventViewModel.Address);
+
+            if (ValidOperation())
+            {
+                string url = Url.Action("GetAddress", "Events", new { id = eventViewModel.Id });
+                return Json(new { success = true, url = url });
+            }
+            return PartialView("_UpdateAddress", eventViewModel);
+        }
+        public IActionResult GetAddress(Guid id)
+        {
+            return PartialView("_DetailsAddress", _eventAppService.GetById(id));
+        }
+        private bool ValidateAuthorEvent(EventViewModel eventViewModel)
+        {
+            return eventViewModel.HostId != HostId;
         }
     }
 }
